% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxtv.R
\name{cindex_tv}
\alias{cindex_tv}
\title{Compute the concordance index of the fit}
\usage{
cindex_tv(phe, tv_list, ti_vars, B)
}
\arguments{
\item{B}{The fitted paraneter, all the other variables are the same as above}
}
\value{
The C-index
}
\description{
Compute the concordance index of the fit
}
