% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxtv.R
\name{coxtv}
\alias{coxtv}
\title{Fit a proportional hazard model with time-dependent covariates}
\usage{
coxtv(phe, tv_list, ti_vars, lambda, B0 = NULL, p.fac = NULL, info = NULL)
}
\arguments{
\item{phe}{a data frame or data table with the following columns:
y: time-to-event;
status: a binary variable, 1 if event  happened, 0 if censored;
time independent covariates specified in ti_vars
ID: the ID of each individual}

\item{tv_list}{a list of data frame or data table, each corresponding to
one time-varying covariate. The elements of this list must
have the following columns:
ID: same ID in phe
time: time when the measurement taken relative to t0, where
time-to-event is defined as t0 -> t1
value: the value of the measurement
Note that the user is responsible to ensure that each ID must
have a measurement before the first event time}

\item{ti_vars}{the names of time-independent predictors}

\item{lambda}{a vector of regularization parameters}

\item{B0}{the initial parameter, zero if not provided}

\item{p.fac}{penalty factor for each variable, uniform if not provided}
}
\value{
The lasso solution
}
\description{
Fit a proportional hazard model with time-dependent covariates
}
