% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxtv.R
\name{KM_curve_tv}
\alias{KM_curve_tv}
\title{Draw a Kaplan Meier curve based on a time-varying covariate}
\usage{
KM_curve_tv(phe, tv, ngroup = 2)
}
\arguments{
\item{tv}{The time varying covariate. Must satisfy the same condition on the
elements of tv_list above}
}
\value{
The KM-curve
}
\description{
Draw a Kaplan Meier curve based on a time-varying covariate
}
